+++ controllers/steps/install/interfaces.go
@@ -217,3 +217,7 @@ type expositionManager interface {
 	EnsureExposition(ctx context.Context, doguResource *v2.Dogu, doguService *coreV1.Service) error
 	RemoveExposition(ctx context.Context, doguName cescommons.SimpleName) error
 }
+
+type warpMenuEntryManager interface {
+	EnsureWarpMenuEntry(ctx context.Context, doguResource *v2.Dogu) error
+}